// src/app/settings/alert-configuration/store/alert-configuration.types.ts
export enum AlertConfigurationActionTypes {
  LOAD_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Load Alert Configuration Request",
  LOAD_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Load Alert Configuration Failure",
  LOAD_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Load Alert Configuration Success",

  UPDATE_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Update Alert Configuration Request",
  UPDATE_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Update Alert Configuration Failure",
  UPDATE_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Update Alert Configuration Success",

  RESET_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Reset Alert Configuration Request",
  RESET_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Reset Alert Configuration Failure",
  RESET_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Reset Alert Configuration Success",

  LOAD_DESTINATION_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Load Destination Alert Configuration Success",
  LOAD_DESTINATION_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Load Destination Alert Configuration Request",
  LOAD_DESTINATION_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Load Destination Alert Configuration Failure",

  LOAD_GROUP_DESTINATION_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Load Group Destination Alert Configuration Success",
  LOAD_GROUP_DESTINATION_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Load Group Destination Alert Configuration Request",
  LOAD_GROUP_DESTINATION_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Load Group Destination Alert Configuration Failure",

  LOAD_GROUP_ALERT_CONFIGURATION_REQUEST = "[ALERTSCONFIGURATION] Load Group Alert Configuration Request",
  LOAD_GROUP_ALERT_CONFIGURATION_FAILURE = "[ALERTSCONFIGURATION] Load Group Alert Configuration Failure",
  LOAD_GROUP_ALERT_CONFIGURATION_SUCCESS = "[ALERTSCONFIGURATION] Load Group Alert Configuration Success",

  LOAD_ARRHYTHMIA_ALERT_CONFIGURATION_REQUEST = "[ARRHYTHMIAALERTSCONFIGURATION] Load Arrhythmia Alert Configuration Request",
  LOAD_ARRHYTHMIA_ALERT_CONFIGURATION_FAILURE = "[ARRHYTHMIAALERTSCONFIGURATION] Load Arrhythmia Alert Configuration Failure",
  LOAD_ARRHYTHMIA_ALERT_CONFIGURATION_SUCCESS = "[ARRHYTHMIAALERTSCONFIGURATION] Load Arrhythmia Alert Configuration Success",

  SET_GROUP_ALERT_ID = "[ALERTSCONFIGURATION] Set Group Alert Id",

  CF_CONATCT_LIST_REQUEST = "[ALERTSCONFIGURATION] CF Contact list Request",
  CF_CONATCT_LIST_FAILURE = "[ALERTSCONFIGURATION] CF Contact list Failure",
  CF_CONATCT_LIST_SUCCESS = "[ALERTSCONFIGURATION] CF Contact list Success",
}